[tests] get_all_ipv4() returns even if SSH fails.
To test this, a new mocked virtual machine class was introduced, on
which all methods are mocked, except those in the base VM class.
Creating a mocked SSH channel is forced to produce an exception and the
function get_all_ipv4() of the newly introduced mock is exercised this
way.